+++ images/tiler-cache/Dockerfile
@@ -14,4 +14,5 @@ RUN pip install --no-cache-dir --upgrade -r requirements.txt
 ENV PYTHONPATH="/app"
 COPY *.sh .
 COPY *.py .
+COPY utils/ .
 CMD ["python", "sqs_processor.py"]
